+++ common/client/fragments/project-fragment/src/helpers/filter-projects.helper.ts
@@ -1,5 +1,5 @@
 import { Project }    from '@shared/data'
-import { isIncludes } from '@shared/helpers'
+import { isIncludes } from '@shared/utils'
 
 export const filterProjects = (
   projects: Project[],
