+++ common/client/fragments/team-invite-fragment/package.json
@@ -4,8 +4,8 @@
   "main": "src/index.ts",
   "dependencies": {
     "@shared/data": "workspace:0.0.1",
-    "@shared/helpers": "workspace:0.0.1",
     "@shared/routes": "workspace:0.0.1",
+    "@shared/utils": "workspace:0.0.1",
     "@ui/avatar": "workspace:0.0.1",
     "@ui/button": "workspace:0.0.1",
     "@ui/condition": "workspace:0.0.1",
