+++ common/client/fragments/team-invite-fragment/src/team-invite.component.tsx
@@ -16,8 +16,8 @@ import { Tag }              from '@ui/tag'
 import { Text }             from '@ui/text'
 import { Title }            from '@ui/title'
 import { WrapperWhite }     from '@ui/wrapper'
-import { getUniqueTags }    from '@shared/helpers'
 import { routes }           from '@shared/routes'
+import { getUniqueTags }    from '@shared/utils'
 
 import { useGetUsers }      from './data'
 import { filterUsers }      from './helpers'
