+++ common/client/fragments/user-teams-fragment/package.json
@@ -4,7 +4,7 @@
   "main": "src/index.ts",
   "dependencies": {
     "@shared/data": "workspace:0.0.1",
-    "@shared/helpers": "workspace:0.0.1",
+    "@shared/utils": "workspace:0.0.1",
     "@ui/button": "workspace:0.0.1",
     "@ui/card": "workspace:0.0.1",
     "@ui/condition": "workspace:0.0.1",
