+++ features/identity/client/fragments/auth-forms/package.json
@@ -5,7 +5,7 @@
   "main": "src/index.ts",
   "dependencies": {
     "@identity/messages-fragment": "workspace:0.0.1",
-    "@shared/helpers": "workspace:0.0.1",
+    "@shared/utils": "workspace:0.0.1",
     "@ui/background": "workspace:0.0.1",
     "@ui/button": "workspace:0.0.1",
     "@ui/condition": "workspace:0.0.1",
