+++ features/identity/client/fragments/auth-forms/src/recovery-new-password/recovery-new-password-form.component.tsx
@@ -19,7 +19,7 @@ import { Row }               from '@ui/layout'
 import { Column }            from '@ui/layout'
 import { Layout }            from '@ui/layout'
 import { Text }              from '@ui/text'
-import { checkPassword }     from '@shared/helpers'
+import { checkPassword }     from '@shared/utils'
 
 export const RecoveryNewPasswordForm = () => {
   const [showPassword, setShowPassword] = useState<boolean>(false)
