+++ shared/utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@shared/helpers",
+  "name": "@shared/utils",
   "version": "0.0.1",
   "license": "BSD-3-Clause",
   "main": "src/index.ts"
