+++ shared/utils/src/helpers/check-password.ts
