+++ ui/card/package.json
@@ -4,7 +4,6 @@
   "main": "src/index.ts",
   "dependencies": {
     "@shared/data": "workspace:0.0.1",
-    "@shared/helpers": "workspace:0.0.1",
     "@shared/routes": "workspace:0.0.1",
     "@ui/avatar": "workspace:0.0.1",
     "@ui/background": "workspace:0.0.1",
